feat(preprocess): allows cdo commands to be run after selection, and allows for send_as
CDO commands can be given as a list as part of the sending specifications,
which are run one after the other, always using the last output file as the
next input file. Support for CDO with multiple inputs is not supported.

Furthermore, send_as is supported as a flag for each variable, which will be
used to rename the variables before the final merge is made.